Add playGame function and create for loop to play
Wrap the playGame function around the whole playRound function to
access everything, move important variables to the playGame function.
Create for loop to go through the playRound function 5 times and log the
round and score each round by having the score variables and the const
variables that give input for playRound in the for loop.
In separate if statement after the 5 rounds use to find out who the
winner is and show it in console with scores of both player and
computer. Start game with playGame to call the function