Rewrite to use buttons instead of prompt
First use defer in html to make it call later and be able to work.
Remove all player input and replace with buttons. Append the buttons to
the HTML and add EventListener to each to link it to the inputs of Rock,
Paper and Scissor. Each will then call the playRound function with their
given value. Add div and append, it displays the result. Add function to
update the Score and display it on the HTML. Add function that checks
and then states who the winner is and displays it on the HTML with
append. Lastely add function to disable all button after the game is
done played for style reasons